+++ src/redux/actions/experimentSettings/backendStatus/updateBackendStatus.js
@@ -1,6 +1,6 @@
 import {
   EXPERIMENT_SETTINGS_BACKEND_STATUS_LOADED,
-} from '../../actionTypes/experimentSettings';
+} from '../../../actionTypes/experimentSettings';
 
 const updateBackendStatus = (status) => async (dispatch) => {
   dispatch({
