+++ src/redux/actions/experimentSettings/processingConfig/setQCStepEnabled.js
@@ -1,4 +1,4 @@
-import { EXPERIMENT_SETTINGS_SET_QC_STEP_ENABLED } from '../../actionTypes/experimentSettings';
+import { EXPERIMENT_SETTINGS_SET_QC_STEP_ENABLED } from '../../../actionTypes/experimentSettings';
 
 const setQCStepEnabled = (step, enabled) => (dispatch) => {
   dispatch({
