+++ src/redux/actions/experimentSettings/processingConfig/updateSampleFilterSettings.js
@@ -1,4 +1,4 @@
-import { EXPERIMENT_SETTINGS_SAMPLE_FILTER_UPDATE } from '../../actionTypes/experimentSettings';
+import { EXPERIMENT_SETTINGS_SAMPLE_FILTER_UPDATE } from '../../../actionTypes/experimentSettings';
 
 const updateSampleFilterSettings = (step, sampleId, diff) => (dispatch) => {
   dispatch({
