+++ src/redux/reducers/experimentSettings/backendStatus/backendStatusLoaded.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 import produce, { original } from 'immer';
 
-import initialState from './initialState';
+import initialState from '../initialState';
 
 const backendStatusLoaded = produce((draft, action) => {
   const { status } = action.payload;
