+++ src/redux/reducers/experimentSettings/backendStatus/backendStatusLoading.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 import produce from 'immer';
 
-import initialState from './initialState';
+import initialState from '../initialState';
 
 const backendStatusLoading = produce((draft) => {
   draft.backendStatus.loading = true;
