+++ src/redux/reducers/experimentSettings/processingConfig/copyFilterSettingsToAllSamples.js
