+++ src/redux/reducers/experimentSettings/processingConfig/setQCStepEnabled.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 import produce from 'immer';
 
-import initialState from './initialState';
+import initialState from '../initialState';
 
 const setQCStepEnabled = produce((draft, action) => {
   const { step, enabled } = action.payload;
